pyproject.toml
[tool.poetry]
name = "django3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
ipython = "^8.22.2"
pytils = "^0.4.1"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
